Fixed 3D multiplexing

+++ encode3dBD.sh
@@ -49,35 +49,37 @@ h264File="$outdir/tmp.$workingFile.264"
 outfile="$outdir/`basename "$mkv"`"
 out3d="$outdir/tmp.3d.$workingFile.mkv"
 
-echo mkvextract tracks "$workingFile" "$streamId:$h264File"
-test -s "$h264File" || mkvextract tracks "$workingFile" "$streamId:$h264File"
+if [ ! -f "$out3d" ]; then
+    echo mkvextract tracks "$workingFile" "$streamId:$h264File"
+    test -s "$h264File" || mkvextract tracks "$workingFile" "$streamId:$h264File"
 
-filter="scale=$streamWidth:$streamHeight"
-test "$crop" && filter="$filter,$crop"
-filter=( '-vf' "$filter" )
+    filter="scale=$streamWidth:$streamHeight"
+    test "$crop" && filter="$filter,$crop"
+    filter=( '-vf' "$filter" )
+
+    ffmpegCmd=(
+        ffmpeg -y -f rawvideo
+        -s:v "$(( $streamWidth * 2 ))x$streamHeight"
+        -r "$streamFps"
+        -i -
+        -c:v libx264
+        "${filter[@]}"
+        -preset medium -tune film
+        -b-pyramid normal -partitions p8x8,b8x8,i4x4
+        "file:$out3d"
+    )
+    wineCmd=( wine FRIMDecode -i:mvc "$h264File" -o - -sbs )
+
+    "${wineCmd[@]}" | "${ffmpegCmd[@]}"
+
+    newLength=`du -m "$out3d" | cut -f 1`
+
+    # only remove h264 file if result is larger than 20m
+    test "$newLength" -gt 20 && rm "$h264File" || exit 1
+fi
 
 ffmpegCmd=(
-    ffmpeg -y -f rawvideo
-    -s:v "$(( $streamWidth * 2 ))x$streamHeight"
-    -r "$streamFps"
-    -i -
-    -c:v libx264
-    "${filter[@]}"
-    -preset medium -tune film
-    -b-pyramid normal -partitions p8x8,b8x8,i4x4
-    "$out3d"
-)
-wineCmd=( wine FRIMDecode -i:mvc "$h264File" -o - -sbs )
-
-"${wineCmd[@]}" | "${ffmpegCmd[@]}"
-
-newLength=`du -m "$out3d" | cut -f 1`
-
-# only remove h264 file if result is larger than 20m
-test "$newLength" -gt 20 && rm "$h264File" || exit 1
-
-ffmpegCmd=(
-    ffmpeg -i "file:$out3d" -i "$mkv" -map 0:v -map 1 -map -1:v
+    ffmpeg -i "file:$out3d" -i "file:$mkv" -map 0:v -map 1 -map -1:v
     -c copy
     "file:$outfile"
 )
